+++ aries_vcx/src/libindy/credentials/mod.rs
@@ -1,1 +1,3 @@
 pub mod encoding;
+pub mod holder;
+pub mod issuer;
