+++ aries_vcx/src/libindy/ledger/transactions.rs
@@ -3,18 +3,18 @@ use std::collections::HashMap;
 use futures::future::TryFutureExt;
 use vdrtools::cache;
 use vdrtools::ledger;
-use vdrtools_sys::{WalletHandle, PoolHandle};
+use vdrtools_sys::{PoolHandle, WalletHandle};
 use serde_json;
 
 use messages::did_doc::service_aries::AriesService;
 use crate::error::prelude::*;
 use crate::global::settings;
 use crate::libindy::utils::mocks::pool_mocks::PoolMocks;
-use crate::libindy::signus::create_and_store_my_did;
+use crate::libindy::keys::create_and_store_my_did;
 use messages::connection::did::Did;
 use messages::connection::invite::Invitation;
 use crate::utils;
-use crate::utils::constants::SUBMIT_SCHEMA_RESPONSE;
+use crate::utils::constants::{CRED_DEF_ID, CRED_DEF_JSON, CRED_DEF_REQ, rev_def_json, REV_REG_DELTA_JSON, REV_REG_ID, REV_REG_JSON, REVOC_REG_TYPE, SCHEMA_ID, SCHEMA_JSON, SCHEMA_TXN, SUBMIT_SCHEMA_RESPONSE};
 use crate::utils::random::generate_random_did;
 use messages::did_doc::service_resolvable::ServiceResolvable;
 use messages::did_doc::DidDoc;
@@ -409,46 +409,6 @@ fn get_data_from_response(resp: &str) -> VcxResult<serde_json::Value> {
         .map_err(|err| VcxError::from_msg(VcxErrorKind::InvalidLedgerResponse, format!("{:?}", err)))
 }
 
-#[cfg(test)]
-#[cfg(feature = "general_test")]
-mod test {
-    use crate::utils::devsetup::*;
-    use messages::a2a::MessageId;
-    use messages::did_doc::test_utils::{_service_endpoint, _recipient_keys, _routing_keys};
-    use messages::connection::invite::test_utils::_pairwise_invitation;
-
-    use super::*;
-
-    #[test]
-    fn test_verify_transaction_can_be_endorsed() {
-        let _setup = SetupDefaults::init();
-
-        // success
-        let transaction = r#"{"reqId":1, "identifier": "EbP4aYNeTHL6q385GuVpRV", "signature": "gkVDhwe2", "endorser": "NcYxiDXkpYi6ov5FcYDi1e"}"#;
-        assert!(_verify_transaction_can_be_endorsed(transaction, "NcYxiDXkpYi6ov5FcYDi1e").is_ok());
-
-        // no author signature
-        let transaction =
-            r#"{"reqId":1, "identifier": "EbP4aYNeTHL6q385GuVpRV", "endorser": "NcYxiDXkpYi6ov5FcYDi1e"}"#;
-        assert!(_verify_transaction_can_be_endorsed(transaction, "NcYxiDXkpYi6ov5FcYDi1e").is_err());
-
-        // different endorser did
-        let transaction =
-            r#"{"reqId":1, "identifier": "EbP4aYNeTHL6q385GuVpRV", "endorser": "NcYxiDXkpYi6ov5FcYDi1e"}"#;
-        assert!(_verify_transaction_can_be_endorsed(transaction, "EbP4aYNeTHL6q385GuVpRV").is_err());
-    }
-
-    #[tokio::test]
-    async fn test_did_doc_from_invitation_works() {
-        let mut did_doc = DidDoc::default();
-        did_doc.set_id(MessageId::id().0);
-        did_doc.set_service_endpoint(_service_endpoint());
-        did_doc.set_recipient_keys(_recipient_keys());
-        did_doc.set_routing_keys(_routing_keys());
-        assert_eq!(did_doc, into_did_doc(0, &Invitation::Pairwise(_pairwise_invitation())).await.unwrap());
-    }
-}
-
 #[derive(Deserialize, Debug)]
 #[serde(rename_all = "camelCase")]
 pub struct Request {
@@ -521,3 +481,367 @@ pub async fn add_new_did(wallet_handle: WalletHandle, pool_handle: PoolHandle, s
         .unwrap();
     (did, verkey)
 }
+
+pub async fn libindy_build_revoc_reg_def_request(submitter_did: &str, rev_reg_def_json: &str) -> VcxResult<String> {
+    if settings::indy_mocks_enabled() {
+        return Ok("".to_string());
+    }
+
+    ledger::build_revoc_reg_def_request(submitter_did, rev_reg_def_json)
+        .await
+        .map_err(VcxError::from)
+}
+
+pub async fn libindy_build_revoc_reg_entry_request(
+    submitter_did: &str,
+    rev_reg_id: &str,
+    rev_def_type: &str,
+    value: &str,
+) -> VcxResult<String> {
+    if settings::indy_mocks_enabled() {
+        return Ok("".to_string());
+    }
+
+    ledger::build_revoc_reg_entry_request(submitter_did, rev_reg_id, rev_def_type, value)
+        .await
+        .map_err(VcxError::from)
+}
+
+pub async fn libindy_build_get_revoc_reg_def_request(submitter_did: &str, rev_reg_id: &str) -> VcxResult<String> {
+    ledger::build_get_revoc_reg_def_request(Some(submitter_did), rev_reg_id)
+        .await
+        .map_err(VcxError::from)
+}
+
+pub async fn libindy_parse_get_revoc_reg_def_response(rev_reg_def_json: &str) -> VcxResult<(String, String)> {
+    ledger::parse_get_revoc_reg_def_response(rev_reg_def_json)
+        .await
+        .map_err(VcxError::from)
+}
+
+pub async fn libindy_build_get_revoc_reg_delta_request(
+    submitter_did: &str,
+    rev_reg_id: &str,
+    from: i64,
+    to: i64,
+) -> VcxResult<String> {
+    ledger::build_get_revoc_reg_delta_request(Some(submitter_did), rev_reg_id, from, to)
+        .await
+        .map_err(VcxError::from)
+}
+
+async fn libindy_build_get_revoc_reg_request(
+    submitter_did: &str,
+    rev_reg_id: &str,
+    timestamp: u64,
+) -> VcxResult<String> {
+    ledger::build_get_revoc_reg_request(Some(submitter_did), rev_reg_id, timestamp as i64)
+        .await
+        .map_err(VcxError::from)
+}
+
+async fn libindy_parse_get_revoc_reg_response(get_cred_def_resp: &str) -> VcxResult<(String, String, u64)> {
+    ledger::parse_get_revoc_reg_response(get_cred_def_resp)
+        .await
+        .map_err(VcxError::from)
+}
+
+async fn libindy_parse_get_cred_def_response(get_rev_reg_resp: &str) -> VcxResult<(String, String)> {
+    ledger::parse_get_cred_def_response(get_rev_reg_resp)
+        .await
+        .map_err(VcxError::from)
+}
+
+pub async fn libindy_parse_get_revoc_reg_delta_response(
+    get_rev_reg_delta_response: &str,
+) -> VcxResult<(String, String, u64)> {
+    ledger::parse_get_revoc_reg_delta_response(get_rev_reg_delta_response)
+        .await
+        .map_err(VcxError::from)
+}
+
+pub async fn build_schema_request(submitter_did: &str, schema: &str) -> VcxResult<String> {
+    trace!("build_schema_request >>> submitter_did: {}, schema: {}", submitter_did, schema);
+
+    if settings::indy_mocks_enabled() {
+        return Ok(SCHEMA_TXN.to_string());
+    }
+
+    let request = libindy_build_schema_request(&submitter_did, schema).await?;
+
+    let request = append_txn_author_agreement_to_request(&request).await?;
+
+    Ok(request)
+}
+
+pub async fn build_rev_reg_request(issuer_did: &str, rev_reg_def_json: &str) -> VcxResult<String> {
+    if settings::indy_mocks_enabled() {
+        debug!("build_rev_reg_request >>> returning mocked value");
+        return Ok("".to_string());
+    }
+
+    let rev_reg_def_req = libindy_build_revoc_reg_def_request(issuer_did, rev_reg_def_json).await?;
+    let rev_reg_def_req = append_txn_author_agreement_to_request(&rev_reg_def_req).await?;
+    Ok(rev_reg_def_req)
+}
+
+pub async fn get_rev_reg_def_json(pool_handle: PoolHandle, rev_reg_id: &str) -> VcxResult<(String, String)> {
+    if settings::indy_mocks_enabled() {
+        debug!("get_rev_reg_def_json >>> returning mocked value");
+        return Ok((REV_REG_ID.to_string(), rev_def_json()));
+    }
+
+    let submitter_did = generate_random_did();
+
+    libindy_build_get_revoc_reg_def_request(&submitter_did, rev_reg_id)
+        .and_then(|req| async move { libindy_submit_request(pool_handle, &req).await })
+        .and_then(|response| async move { libindy_parse_get_revoc_reg_def_response(&response).await })
+        .await
+}
+
+pub async fn build_rev_reg_delta_request(
+    issuer_did: &str,
+    rev_reg_id: &str,
+    rev_reg_entry_json: &str,
+) -> VcxResult<String> {
+    trace!(
+        "build_rev_reg_delta_request >>> issuer_did: {}, rev_reg_id: {}, rev_reg_entry_json: {}",
+        issuer_did,
+        rev_reg_id,
+        rev_reg_entry_json
+    );
+    let request =
+        libindy_build_revoc_reg_entry_request(issuer_did, rev_reg_id, REVOC_REG_TYPE, rev_reg_entry_json).await?;
+    let request = append_txn_author_agreement_to_request(&request).await?;
+    Ok(request)
+}
+
+pub async fn get_rev_reg_delta_json(
+    pool_handle: PoolHandle,
+    rev_reg_id: &str,
+    from: Option<u64>,
+    to: Option<u64>,
+) -> VcxResult<(String, String, u64)> {
+    trace!(
+        "get_rev_reg_delta_json >>> pool_handle: {:?}, rev_reg_id: {}, from: {:?}, to: {:?}",
+        pool_handle,
+        rev_reg_id,
+        from,
+        to
+    );
+    if settings::indy_mocks_enabled() {
+        debug!("get_rev_reg_delta_json >>> returning mocked value");
+        return Ok((REV_REG_ID.to_string(), REV_REG_DELTA_JSON.to_string(), 1));
+    }
+
+    let submitter_did = generate_random_did();
+
+    let from: i64 = if let Some(_from) = from { _from as i64 } else { -1 };
+    let to = if let Some(_to) = to {
+        _to as i64
+    } else {
+        time::get_time().sec
+    };
+
+    libindy_build_get_revoc_reg_delta_request(&submitter_did, rev_reg_id, from, to)
+        .and_then(|req| async move { libindy_submit_request(pool_handle, &req).await })
+        .and_then(|response| async move { libindy_parse_get_revoc_reg_delta_response(&response).await })
+        .await
+}
+
+pub async fn get_rev_reg(pool_handle: PoolHandle, rev_reg_id: &str, timestamp: u64) -> VcxResult<(String, String, u64)> {
+    if settings::indy_mocks_enabled() {
+        return Ok((REV_REG_ID.to_string(), REV_REG_JSON.to_string(), 1));
+    }
+
+    let submitter_did = generate_random_did();
+
+    libindy_build_get_revoc_reg_request(&submitter_did, rev_reg_id, timestamp)
+        .and_then(|req| async move { libindy_submit_request(pool_handle, &req).await })
+        .and_then(|response| async move { libindy_parse_get_revoc_reg_response(&response).await })
+        .await
+}
+
+pub async fn get_cred_def(pool_handle: PoolHandle, issuer_did: Option<&str>, cred_def_id: &str) -> VcxResult<(String, String)> {
+    if settings::indy_mocks_enabled() {
+        return Err(VcxError::from(VcxErrorKind::LibndyError(309)));
+    }
+    libindy_build_get_cred_def_request(issuer_did, cred_def_id)
+        .and_then(|req| async move { libindy_submit_request(pool_handle, &req).await })
+        .and_then(|response| async move { libindy_parse_get_cred_def_response(&response).await })
+        .await
+}
+
+pub async fn is_cred_def_on_ledger(pool_handle: PoolHandle, issuer_did: Option<&str>, cred_def_id: &str) -> VcxResult<bool> {
+    match get_cred_def(pool_handle, issuer_did, cred_def_id).await {
+        Ok(_) => Ok(true),
+        Err(err) if err.kind() == VcxErrorKind::LibndyError(309) => Ok(false),
+        Err(err) => Err(VcxError::from_msg(
+            VcxErrorKind::InvalidLedgerResponse,
+            format!(
+                "Failed to check presence of credential definition id {} on the ledger\nError: {}",
+                cred_def_id, err
+            ),
+        )),
+    }
+}
+
+async fn libindy_build_get_txn_request(submitter_did: Option<&str>, seq_no: i32) -> VcxResult<String> {
+    ledger::build_get_txn_request(submitter_did, None, seq_no)
+        .await
+        .map_err(VcxError::from)
+}
+
+pub async fn build_get_txn_request(submitter_did: Option<&str>, seq_no: i32) -> VcxResult<String> {
+    trace!(
+        "build_get_txn_request >>> submitter_did: {:?}, seq_no: {}",
+        submitter_did,
+        seq_no
+    );
+    let request = libindy_build_get_txn_request(submitter_did, seq_no).await?;
+    let request = append_txn_author_agreement_to_request(&request).await?;
+    Ok(request)
+}
+
+pub async fn get_ledger_txn(
+    wallet_handle: WalletHandle,
+    pool_handle: PoolHandle,
+    submitter_did: Option<&str>,
+    seq_no: i32,
+) -> VcxResult<String> {
+    trace!(
+        "get_ledger_txn >>> submitter_did: {:?}, seq_no: {}",
+        submitter_did,
+        seq_no
+    );
+    let req = build_get_txn_request(submitter_did, seq_no).await?;
+    let res = if let Some(submitter_did) = submitter_did {
+        libindy_sign_and_submit_request(wallet_handle, pool_handle, submitter_did, &req).await?
+    } else {
+        libindy_submit_request(pool_handle, &req).await?
+    };
+    check_response(&res)?;
+    Ok(res)
+}
+
+pub fn _check_schema_response(response: &str) -> VcxResult<()> {
+    // TODO: saved backwardcampatibilyty but actually we can better handle response
+    match parse_response(response)? {
+        Response::Reply(_) => Ok(()),
+        Response::Reject(reject) => Err(VcxError::from_msg(
+            VcxErrorKind::DuplicationSchema,
+            format!("{:?}", reject),
+        )),
+        Response::ReqNACK(reqnack) => Err(VcxError::from_msg(
+            VcxErrorKind::UnknownSchemaRejection,
+            format!("{:?}", reqnack),
+        )),
+    }
+}
+
+pub(in crate::libindy) fn check_response(response: &str) -> VcxResult<()> {
+    match parse_response(response)? {
+        Response::Reply(_) => Ok(()),
+        Response::Reject(res) | Response::ReqNACK(res) => Err(VcxError::from_msg(
+            VcxErrorKind::InvalidLedgerResponse,
+            format!("{:?}", res),
+        )),
+    }
+}
+
+pub async fn get_schema_json(wallet_handle: WalletHandle, pool_handle: PoolHandle, schema_id: &str) -> VcxResult<(String, String)> {
+    trace!("get_schema_json >>> schema_id: {}", schema_id);
+    if settings::indy_mocks_enabled() {
+        return Ok((SCHEMA_ID.to_string(), SCHEMA_JSON.to_string()));
+    }
+
+    let submitter_did = generate_random_did();
+
+    let schema_json = libindy_get_schema(wallet_handle, pool_handle, &submitter_did, schema_id).await?;
+
+    Ok((schema_id.to_string(), schema_json))
+}
+
+pub async fn build_cred_def_request(issuer_did: &str, cred_def_json: &str) -> VcxResult<String> {
+    if settings::indy_mocks_enabled() {
+        return Ok(CRED_DEF_REQ.to_string());
+    }
+
+    let cred_def_req = libindy_build_create_credential_def_txn(issuer_did, cred_def_json).await?;
+
+    let cred_def_req = append_txn_author_agreement_to_request(&cred_def_req).await?;
+
+    Ok(cred_def_req)
+}
+
+pub async fn get_cred_def_json(wallet_handle: WalletHandle, pool_handle: PoolHandle, cred_def_id: &str) -> VcxResult<(String, String)> {
+    if settings::indy_mocks_enabled() {
+        debug!("get_cred_def_json >>> returning mocked value");
+        return Ok((CRED_DEF_ID.to_string(), CRED_DEF_JSON.to_string()));
+    }
+
+    let cred_def_json = libindy_get_cred_def(wallet_handle, pool_handle, cred_def_id).await?;
+
+    Ok((cred_def_id.to_string(), cred_def_json))
+}
+
+
+#[cfg(test)]
+#[cfg(feature = "general_test")]
+mod test {
+    use crate::utils::devsetup::*;
+    use messages::a2a::MessageId;
+    use messages::did_doc::test_utils::{_recipient_keys, _routing_keys, _service_endpoint};
+    use messages::connection::invite::test_utils::_pairwise_invitation;
+
+    use super::*;
+
+    #[test]
+    fn test_verify_transaction_can_be_endorsed() {
+        let _setup = SetupDefaults::init();
+
+        // success
+        let transaction = r#"{"reqId":1, "identifier": "EbP4aYNeTHL6q385GuVpRV", "signature": "gkVDhwe2", "endorser": "NcYxiDXkpYi6ov5FcYDi1e"}"#;
+        assert!(_verify_transaction_can_be_endorsed(transaction, "NcYxiDXkpYi6ov5FcYDi1e").is_ok());
+
+        // no author signature
+        let transaction =
+            r#"{"reqId":1, "identifier": "EbP4aYNeTHL6q385GuVpRV", "endorser": "NcYxiDXkpYi6ov5FcYDi1e"}"#;
+        assert!(_verify_transaction_can_be_endorsed(transaction, "NcYxiDXkpYi6ov5FcYDi1e").is_err());
+
+        // different endorser did
+        let transaction =
+            r#"{"reqId":1, "identifier": "EbP4aYNeTHL6q385GuVpRV", "endorser": "NcYxiDXkpYi6ov5FcYDi1e"}"#;
+        assert!(_verify_transaction_can_be_endorsed(transaction, "EbP4aYNeTHL6q385GuVpRV").is_err());
+    }
+
+    #[tokio::test]
+    async fn test_did_doc_from_invitation_works() {
+        let mut did_doc = DidDoc::default();
+        did_doc.set_id(MessageId::id().0);
+        did_doc.set_service_endpoint(_service_endpoint());
+        did_doc.set_recipient_keys(_recipient_keys());
+        did_doc.set_routing_keys(_routing_keys());
+        assert_eq!(did_doc, into_did_doc(0, &Invitation::Pairwise(_pairwise_invitation())).await.unwrap());
+    }
+}
+
+#[cfg(test)]
+#[cfg(feature = "pool_tests")]
+pub mod integration_tests {
+    use crate::libindy::ledger::transactions::get_ledger_txn;
+    use crate::utils::devsetup::SetupWalletPool;
+
+    #[tokio::test]
+    async fn test_get_txn() {
+        let setup = SetupWalletPool::init().await;
+        get_ledger_txn(setup.wallet_handle, setup.pool_handle, None, 0).await.unwrap_err();
+        let txn = get_ledger_txn(setup.wallet_handle, setup.pool_handle, None, 1).await;
+        assert!(txn.is_ok());
+
+        get_ledger_txn(setup.wallet_handle, setup.pool_handle, Some(&setup.institution_did), 0).await.unwrap_err();
+        let txn = get_ledger_txn(setup.wallet_handle, setup.pool_handle, Some(&setup.institution_did), 1).await;
+        assert!(txn.is_ok());
+    }
+}
+
