+++ aries_vcx/src/libindy/mod.rs
@@ -3,9 +3,11 @@ pub mod credentials;
 pub mod proofs;
 pub mod utils;
 pub mod wallet;
-pub mod signus;
-pub mod crypto;
+pub mod keys;
+pub mod signing;
 pub mod wallet_non_secrets;
 pub mod anoncreds;
 pub mod ledger;
 pub mod primitives;
+#[cfg(feature = "test_utils")]
+pub mod test_utils;
