+++ aries_vcx/src/libindy/signing.rs
