+++ aries_vcx/src/protocols/connection/pairwise_info.rs
@@ -1,7 +1,7 @@
 use vdrtools_sys::WalletHandle;
 
 use crate::error::VcxResult;
-use crate::libindy::signus::create_and_store_my_did;
+use crate::libindy::keys::create_and_store_my_did;
 
 #[derive(Debug, Clone, Serialize, Deserialize, PartialEq)]
 pub struct PairwiseInfo {
