+++ aries_vcx/tests/test_creds_proofs.rs
@@ -11,8 +11,8 @@ pub mod utils;
 #[cfg(feature = "pool_tests")]
 mod integration_tests {
     use aries_vcx::handlers::proof_presentation::prover::Prover;
-    use aries_vcx::libindy::anoncreds::get_cred_def_json;
-    use aries_vcx::libindy::anoncreds::test_utils::{
+    use aries_vcx::libindy::ledger::transactions::get_cred_def_json;
+    use aries_vcx::libindy::test_utils::{
         create_and_store_credential, create_and_store_nonrevocable_credential,
         create_and_store_nonrevocable_credential_def, create_indy_proof,
     };
@@ -321,7 +321,7 @@ mod tests {
     use aries_vcx::handlers::proof_presentation::prover::Prover;
     use aries_vcx::handlers::proof_presentation::verifier::Verifier;
     use aries_vcx::libindy;
-    use aries_vcx::libindy::anoncreds::test_utils::create_and_store_nonrevocable_credential_def;
+    use aries_vcx::libindy::test_utils::create_and_store_nonrevocable_credential_def;
     use aries_vcx::libindy::ledger::pool::test_utils::{delete_test_pool, open_test_pool};
     use aries_vcx::messages::issuance::credential_offer::CredentialOffer;
     use aries_vcx::messages::proof_presentation::presentation_request::PresentationRequest;
