+++ libvcx/src/api_lib/api_c/schema.rs
@@ -580,7 +580,7 @@ mod tests {
 
     use aries_vcx::libindy::primitives::credential_definition::PublicEntityStateType;
     use aries_vcx::libindy::primitives::credential_schema::Schema;
-    use aries_vcx::libindy::anoncreds::test_utils::create_and_write_test_schema;
+    use aries_vcx::libindy::test_utils::create_and_write_test_schema;
     use aries_vcx::utils;
     use aries_vcx::utils::constants::{
         DEFAULT_SCHEMA_ATTRS, DEFAULT_SCHEMA_ID, DEFAULT_SCHEMA_NAME, SCHEMA_ID, SCHEMA_WITH_VERSION,
